+++ wswon/AndroidArchitectureStudy/app/src/main/java/com/architecture/study/data/source/CoinRemoteDataSource.kt
@@ -1,0 +1,44 @@
+package com.architecture.study.data.source
+
+import com.architecture.study.network.model.MarketResponse
+import com.architecture.study.network.model.TickerResponse
+import com.architecture.study.network.request.UpbitRequest
+import com.architecture.study.network.request.UpbitRequestListener
+
+object CoinRemoteDataSource : CoinDataSource {
+
+    override fun getMarketList(listener: UpbitRequestListener<MarketResponse>) {
+        UpbitRequest().getMarketList(object : UpbitRequestListener<MarketResponse> {
+            override fun onSucess(dataList: List<MarketResponse>) {
+                listener.onSucess(dataList)
+            }
+
+            override fun onEmpty(str: String) {
+                listener.onEmpty(str)
+            }
+
+            override fun onFailure(str: String) {
+                listener.onFailure(str)
+            }
+        })
+
+    }
+
+    override fun getTickerList(marketNames: String, listener: UpbitRequestListener<TickerResponse>) {
+        UpbitRequest().getTickerList(marketNames, object :
+            UpbitRequestListener<TickerResponse> {
+            override fun onSucess(dataList: List<TickerResponse>) {
+                listener.onSucess(dataList)
+            }
+
+            override fun onEmpty(str: String) {
+                listener.onEmpty(str)
+            }
+
+            override fun onFailure(str: String) {
+                listener.onFailure(str)
+            }
+        })
+    }
+
+}
